backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlmodel==0.0.22
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
redis==5.0.8
celery==5.4.0
sse-starlette==2.1.3
boto3==1.34.162
httpx==0.27.2


